Added a function for DocArrayInMemorySearch
